Refactoring index.ts. Seperating routes + implemenations
